Reorder Economy and Bank cogs, moving stuff from economy that should be in bank
